change input of Tstep to nTemps

+++ projects/Project4/cpp/tools.h
@@ -7,6 +7,6 @@
 using namespace std;
 
 int periodic(int i, int limit, int add);
-int readData(string filename, int& NMC, double & Tstart, double &Tstop, double &Tstep, int &L, bool & timeit, bool & save_to_file, bool & orderedSpinConfig);
+int readData(string filename, int& NMC, double & Tstart, double &Tstop, int &Tstep, int &L, bool & timeit, bool & save_to_file, bool & orderedSpinConfig);
 
 #endif // TOOLS_H
